requirements.txt
alabaster==0.7.12
asgiref==3.2.7
autopep8==1.5.6
Babel==2.9.0
beautifulsoup4==4.9.1
bleach==3.2.1
boto==2.49.0
boto3==1.14.60
botocore==1.17.60
Brotli==1.0.7
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==2.9.2
cssselect==1.1.0
defusedxml==0.7.0rc1
diff-match-patch==20200713
Django==3.0.7
django-allauth==0.42.0
django-cachalot==2.4.2
django-compression-middleware==0.4.1
django-countries==6.1.2
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-easy-audit==1.3.0
django-edit-suggestion==1.40
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.3.0
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-mailer==2.1
django-markdownify==0.9.0
django-mptt==0.12.0
django-recaptcha==2.0.6
django-redis==4.12.1
django-url-alias
django-user-agents==0.4.0
django-versatileimagefield==2.0
djangorestframework==3.11.0
docker==4.3.1
docutils==0.15.2
elasticsearch==7.0.0
Faker==4.1.1
feedfinder2==0.0.4
feedparser==6.0.0
futures==3.1.1
gensim==3.8.3
google-api-core==1.20.1
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
goslate==1.5.1
gprof2dot==2021.2.21
grpcio==1.29.0
httplib2==0.18.1
huey==2.3.0
idna==2.9
imagesize==1.2.0
jeepney==0.6.0
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
keyring==21.5.0
lockfile==0.12.2
lxml==4.5.2
Markdown==3.3.4
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
pkginfo==1.6.1
polib==1.1.0
protobuf==3.12.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.18
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
readme-renderer==28.0
redis==3.5.3
regex==2020.7.14
requests==2.23.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
s3transfer==0.3.3
scipy==1.5.2
SecretStorage==3.2.0
sentry-sdk==0.17.2
sgmllib3k==1.0.0
six==1.15.0
smart-open==2.1.1
snowballstemmer==2.0.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
soupsieve==2.0.1
Sphinx==3.3.0
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
text-unidecode==1.3
tinysegmenter==0.3
tldextract==2.2.3
toml==0.10.2
tqdm==4.49.0
twine==3.2.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.9
user-agents==2.2.0
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
zstandard==0.14.0
